New Data struct for Fissures
The old data struct doesn't work for crossing fissures, I have to change
the principal matrix. Now there are 3 matrix the first is for empty,
breakable and unbreakable field, the second is for fissures and the
third is for water. Data Struct for fissures works. I have to change the
water data struct